+++ examples/project/dialog_box/test_2_input_field.py
@@ -63,3 +63,17 @@ def test_checkbox():
     response = run_dialog_box(dbx)
     set_message(f"Selected item {response}")
     assert response == '["one"]', "The answer is not correct"
+
+
+@pytest.mark.case_name("Image")
+def test_image():
+    dbx = DialogBox(
+        dialog_text=f"Test image",
+        title_bar="Image example",
+        widget=DialogBoxWidget(
+            DialogBoxWidgetType.IMAGE,
+            {"imageUrl": "/text_input_dialog_box.jpg"},
+        ),
+    )
+    response = run_dialog_box(dbx)
+    assert response is None